messaging_app/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
mysqlclient==2.2.0
pytest==7.4.3
pytest-django==4.7.0
flake8==6.1.0
coverage==7.3.2
python-decouple==3.8
gunicorn==21.2.0 